Issue #2247 Ensure random unused ssl port for osgi tests
